+++ zarrs/src/array/array_ops/array_update_ops_array.rs
@@ -19,13 +19,13 @@ impl<TStorage: ?Sized + ReadableWritableStorageTraits + 'static> ArrayUpdateOps
     pub fn store_chunk_subset<'a, T: IntoArrayBytes<'a>>(
         &self,
         chunk_indices: &[u64],
-        chunk_subset: &dyn ArraySubsetTraits,
-        chunk_subset_data: T,
+        indexer: &dyn Indexer,
+        indexer_data: T,
     ) -> Result<(), ArrayError> {
         self.store_chunk_subset_with_options(
             chunk_indices,
-            chunk_subset,
-            chunk_subset_data,
+            indexer,
+            indexer_data,
             self.codec_options(),
         )
     }
@@ -149,60 +149,59 @@ impl<TStorage: ?Sized + ReadableWritableStorageTraits + 'static> Array<TStorage>
     pub(in crate::array) fn store_chunk_subset_with_options<'a, T: IntoArrayBytes<'a>>(
         &self,
         chunk_indices: &[u64],
-        chunk_subset: &dyn ArraySubsetTraits,
-        chunk_subset_data: T,
+        indexer: &dyn Indexer,
+        indexer_data: T,
         options: &CodecOptions,
     ) -> Result<(), ArrayError> {
         let chunk_shape = self
             .chunk_grid()
             .chunk_shape_u64(chunk_indices)?
             .ok_or_else(|| ArrayError::InvalidChunkGridIndicesError(chunk_indices.to_vec()))?;
-        if std::iter::zip(chunk_subset.end_exc(), &chunk_shape)
-            .any(|(end_exc, shape)| end_exc > *shape)
-        {
-            return Err(ArrayError::InvalidChunkSubset(
-                chunk_subset.to_array_subset(),
-                chunk_indices.to_vec(),
-                chunk_shape,
-            ));
+        if let Some(chunk_subset) = indexer.as_array_subset() {
+            if !chunk_subset.inbounds_shape(&chunk_shape) {
+                return Err(ArrayError::InvalidChunkSubset(
+                    chunk_subset.to_array_subset(),
+                    chunk_indices.to_vec(),
+                    chunk_shape,
+                ));
+            }
+            if super::subset_is_whole_chunk(chunk_subset, &chunk_shape) {
+                return self.store_chunk_with_options(chunk_indices, indexer_data, options);
+            }
         }
 
-        if chunk_subset.shape() == chunk_shape && chunk_subset.start().iter().all(|&x| x == 0) {
-            self.store_chunk_with_options(chunk_indices, chunk_subset_data, options)
+        let indexer_bytes = indexer_data.into_array_bytes(self.data_type())?;
+        indexer_bytes.validate(indexer.len(), self.data_type())?;
+
+        // Lock the chunk
+        // let key = self.chunk_key(chunk_indices);
+        // let mutex = self.storage.mutex(&key)?;
+        // let _lock = mutex.lock();
+
+        if options.experimental_partial_encoding()
+            && self.codecs.partial_encoder_capability().partial_encode
+            && self.storage.supports_set_partial()
+        {
+            let partial_encoder = self.partial_encoder_with_options(chunk_indices, options)?;
+            debug_assert!(
+                partial_encoder.supports_partial_encode(),
+                "partial encoder is misrepresenting its capabilities"
+            );
+            Ok(partial_encoder.partial_encode(indexer, &indexer_bytes, options)?)
         } else {
-            let chunk_subset_bytes = chunk_subset_data.into_array_bytes(self.data_type())?;
-            chunk_subset_bytes.validate(chunk_subset.num_elements(), self.data_type())?;
+            let chunk_bytes_old: ArrayBytes<'static> =
+                self.retrieve_chunk_with_options(chunk_indices, options)?;
+            chunk_bytes_old.validate(chunk_shape.iter().product(), self.data_type())?;
 
-            // Lock the chunk
-            // let key = self.chunk_key(chunk_indices);
-            // let mutex = self.storage.mutex(&key)?;
-            // let _lock = mutex.lock();
+            let chunk_bytes_new = update_array_bytes(
+                chunk_bytes_old,
+                &chunk_shape,
+                indexer,
+                &indexer_bytes,
+                self.data_type().size(),
+            )?;
 
-            if options.experimental_partial_encoding()
-                && self.codecs.partial_encoder_capability().partial_encode
-                && self.storage.supports_set_partial()
-            {
-                let partial_encoder = self.partial_encoder_with_options(chunk_indices, options)?;
-                debug_assert!(
-                    partial_encoder.supports_partial_encode(),
-                    "partial encoder is misrepresenting its capabilities"
-                );
-                Ok(partial_encoder.partial_encode(chunk_subset, &chunk_subset_bytes, options)?)
-            } else {
-                let chunk_bytes_old: ArrayBytes<'static> =
-                    self.retrieve_chunk_with_options(chunk_indices, options)?;
-                chunk_bytes_old.validate(chunk_shape.iter().product(), self.data_type())?;
-
-                let chunk_bytes_new = update_array_bytes(
-                    chunk_bytes_old,
-                    &chunk_shape,
-                    chunk_subset,
-                    &chunk_subset_bytes,
-                    self.data_type().size(),
-                )?;
-
-                self.store_chunk_with_options(chunk_indices, chunk_bytes_new, options)
-            }
+            self.store_chunk_with_options(chunk_indices, chunk_bytes_new, options)
         }
     }