+++ zarrs/src/array/chunk_cache/chunk_cache_type/partial_decoder_async.rs
@@ -5,8 +5,7 @@ use crate::array::chunk_cache::{
     AsyncChunkCache, ChunkCacheType, ChunkCacheTypeAsyncPartialDecoder, SealedAsync,
 };
 use crate::array::{
-    Array, ArrayBytes, ArrayError, ArraySubset, ArraySubsetTraits, CodecOptions,
-    chunk_shape_to_array_shape,
+    Array, ArrayBytes, ArrayError, ArraySubset, CodecOptions, Indexer, chunk_shape_to_array_shape,
 };
 use zarrs_codec::AsyncArrayPartialDecoderTraits;
 use zarrs_storage::AsyncReadableStorageTraits;
@@ -70,7 +69,7 @@ impl SealedAsync for ChunkCacheTypeAsyncPartialDecoder {
         cache: &C,
         array: &Array<TStorage>,
         chunk_indices: &[u64],
-        chunk_subset: &dyn ArraySubsetTraits,
+        indexer: &dyn Indexer,
         options: &CodecOptions,
     ) -> Result<Arc<ArrayBytes<'static>>, ArrayError>
     where
@@ -80,7 +79,7 @@ impl SealedAsync for ChunkCacheTypeAsyncPartialDecoder {
         Ok(
             Self::async_partial_decoder(cache, array, chunk_indices, options)
                 .await?
-                .partial_decode(chunk_subset, options)
+                .partial_decode(indexer, options)
                 .await?
                 .into_owned()
                 .into(),